Fixed warning and error parsing
The existing code was using regex to match the line and then used split to spilt the line on spaces. It then assumed that the integer value was the 5th part of the split. This was failing to extract values when building an Unreal project in VS2019.

Regex is now used to both match the line and extract the integer value using regex groups.